rename operator.rs to desugar.rs
